+++ lib/universign/service/transaction.rb
@@ -2,35 +2,33 @@ module Universign
   module Service
     module Transaction
       AVAILABLE_OPTIONS = {
-        custom_id:                  :customId,
-        description:                :description,
+        custom_id: :customId,
+        description: :description,
         handwritten_signature_mode: :handwrittenSignatureMode,
-        certificate_type:           :certificateType,
-        language:                   :language,
-        identification_type:        :identificationType,
-        handwritten_signature:      :handwrittenSignature,
-        profile:                    :profile,
-        final_doc_sent:             :finalDocSent,
-        final_doc_requester_sent:   :finalDocRequesterSent,
-        chaining_mode:              :chainingMode
-      }
+        certificate_type: :certificateType,
+        language: :language,
+        identification_type: :identificationType,
+        handwritten_signature: :handwrittenSignature,
+        profile: :profile,
+        final_doc_sent: :finalDocSent,
+        final_doc_requester_sent: :finalDocRequesterSent,
+        chaining_mode: :chainingMode,
+      }.freeze
 
       DEFAULT_OPTIONS = {
         handwrittenSignatureMode: 1,
-        identificationType:       'sms',
-        language:                 'fr',
-        certificateType:          'simple'
-      }
+        identificationType: "sms",
+        language: "fr",
+        certificateType: "simple",
+      }.freeze
 
       # Get a transaction from Universign
       #
       # @return [Universign::Transaction]
       def get
-        @client = Universign::Client.new
-
         safeguard do
-          result = @client.call('requester.getTransactionInfo', @transaction_id)
-          self.from_data(result)
+          result = Universign::Client.call("requester.getTransactionInfo", @transaction_id)
+          from_data(result)
         end
       end
 
@@ -59,26 +57,21 @@ module Universign
         #
         # @return [Universign::Transaction]
         def create(documents:, signers:, options: {})
-          @client = Universign::Client.new
-
           sign_options = DEFAULT_OPTIONS.merge(
             documents: documents.map(&:params),
-            signers:   signers.map(&:params),
+            signers: signers.map(&:params)
           )
 
           options.each do |key, value|
             known_key = AVAILABLE_OPTIONS[key]
+            raise Universign::UnknownOption, key.to_s unless known_key
 
-            if known_key
-              sign_options[known_key] = value
-            else
-              raise "Unknown Key"
-            end
+            sign_options[known_key] = value
           end
 
           safeguard do
-            result = @client.call("requester.requestTransaction", sign_options)
-            Universign::Transaction.new(result['id'], result['url'])
+            result = Universign::Client.call("requester.requestTransaction", sign_options)
+            Universign::Transaction.new(result["id"], result["url"])
           end
         end
       end