+++ lib/universign/transaction.rb
@@ -1,17 +1,25 @@
+require "uri"
+
 module Universign
   class Transaction
     include Universign::Safeguard
     include Service::Transaction
     include Service::Document
 
-    attr_reader :transaction_id, :url, :data
+    attr_reader :transaction_id
 
-    def initialize(transaction_id = nil, url = nil)
+    def initialize(transaction_id = nil, sign_url = nil)
       @transaction_id = transaction_id
-      @url            = url
-      @data           = {}
+      @sign_url       = sign_url
+    end
 
-      self.get
+    # The transaction info, lazily fetched from Universign on first access.
+    # Transactions built via .create expose their sign_url/signer_id without
+    # triggering this call.
+    #
+    # @return [Hash]
+    def data
+      @data ||= get
     end
 
     def from_data(data)
@@ -22,35 +30,52 @@ module Universign
     #
     # The status of the transaction. The existing statuses are:
     #
-    # | Status      | Description                                                                                      |
-    # |-------------|--------------------------------------------------------------------------------------------------|
-    # | `ready`     | Signers can connect and sign                                                                     |
-    # | `expired`   | The transaction has been requested more than 7 days ago. It will no more be available to signers |
-    # | `canceled`  | A signer has canceled the transaction. Signers will no more be able to connect to the service    |
-    # | `failed`    | An error occured during a signature. The signers won’t be able to connect to the service         |
-    # | `completed` | All signers have successfuly sign, the requester can retrieve the documents                      |
+    # | Status      | Description                                                |
+    # |-------------|------------------------------------------------------------|
+    # | `ready`     | Signers can connect and sign                               |
+    # | `expired`   | Requested more than 7 days ago, no longer available        |
+    # | `canceled`  | A signer has canceled the transaction                      |
+    # | `failed`    | An error occured during a signature                        |
+    # | `completed` | All signers have successfuly signed                        |
     def status
-      data['status']
+      data["status"]
     end
 
-    # @return [Array<String>]
-    def url
-      @url ||= data['signerInfos'].map { |si| si['url'] }
+    # The URL the signer must open to sign (embeddable iframe URL). Always a
+    # String. Available without an API call right after .create.
+    #
+    # @return [String, nil]
+    def sign_url
+      @sign_url ||= data.dig("signerInfos", current_signer || 0, "url")
     end
 
-    # A list of bean containing information about the signers
+    # The signer id, parsed from the sign URL. Universign exposes the id in the
+    # query string on legacy URLs (".../signature/?id=...") and in the fragment
+    # on app.universign.com URLs (".../sig/#/?id=..."), so we read from both.
+    #
+    # @return [String, nil]
+    def signer_id
+      params = URI(sign_url.to_s).then { |uri| uri.query || uri.fragment }
+      return if params.nil?
+
+      URI.decode_www_form(params.sub(%r{\A/?\??}, "")).to_h["id"]
+    end
+
+    # A list of beans containing information about the signers
     # and their progression in the signature process
     #
-    # @return [Array<Universign::Signer]
+    # @return [Array<Universign::SignerInfos>]
     def signers
-      raise 'NotImplementedYet'
+      Array(data["signerInfos"]).map do |signer_info|
+        Universign::SignerInfos.from_data(signer_info)
+      end
     end
 
     # A bean containing information about the requester of a
     # transaction
     # @return
     def initiator
-      data['initiatorInfo']
+      data["initiatorInfo"]
     end
 
     # The index of current signer if the status of transaction
@@ -58,21 +83,21 @@ module Universign
     #
     # @return [Integer]
     def current_signer
-      data['currentSigner']
+      data["currentSigner"]
     end
 
     # The creation date or last relaunch date of this transaction
     #
     # @return [Date]
     def created_at
-      data['creationDate'].to_date
+      data["creationDate"].to_date
     end
 
     # The description of the Transaction
     #
     # @return [String]
     def description
-      data['description']
+      data["description"]
     end
 
     # Whether the transaction was requested with requesting handwritten signature
@@ -80,22 +105,14 @@ module Universign
     #
     # @return [Boolean]
     def each_field
-      data['eachField']
+      data["eachField"]
     end
 
     # Whether the transaction is signed... or not !
     #
     # @return [Boolean]
     def signed?
-      status == 'completed'
-    end
-
-    ########################
-
-    private
-
-    def client
-      Universign::Client.new.client
+      status == "completed"
     end
   end
 end