+++ lib/universign/transaction_signer.rb
@@ -1,6 +1,7 @@
 module Universign
   class TransactionSigner < Signer
-    attr_accessor :phone_number, :signature, :callbacks
+    attr_reader :phone_number
+    attr_accessor :signature, :callbacks
 
     def initialize(options = {})
       super(options)
@@ -56,8 +57,8 @@ module Universign
     #
     # @params [Universign::SignatureField] data
     def signature_field=(data)
-      if !data.instance_of?(Universign::SignatureField)
-        raise 'BadSignatureFieldType' # TODO: create custom Exception
+      unless data.instance_of?(Universign::SignatureField)
+        raise Universign::InvalidSignatureField
       end
 
       @signature_field        = data
@@ -68,7 +69,7 @@ module Universign
     # signature
     #
     # @params [Date] data
-    def birtdate=(data)
+    def birthdate=(data)
       @birthdate         = data
       params[:birthDate] = data
     end
@@ -93,12 +94,12 @@ module Universign
 
     # Which authentication type will be used when a signer will attempt to sign.
     #
-    # The available values are :
-    # |   Type  |                                                                                    Description                                                                                   |
-    # |:-------:|:--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------:|
-    # |  `none` |                                                           The signer won’t be asked an authentication code when signing                                                          |
-    # | `email` |       The signer will be sent a authentication code by e-mail. Using this option implies that this signer has a valid email property set, otherwise, an exception is thrown      |
-    # |  `sms`  | The signer will be sent a authentication code by sms. Using this option implies that this signer has a valid `phone_number` property set, in other cases, an exception is thrown |
+    # The available values are:
+    # - `none`: the signer won't be asked an authentication code when signing
+    # - `email`: an authentication code is sent by e-mail (requires a valid
+    #   `email`, otherwise an exception is thrown)
+    # - `sms`: an authentication code is sent by SMS (requires a valid
+    #   `phone_number`, otherwise an exception is thrown)
     def identification_type=(data)
       @identification_type        = data
       params[:identificationType] = data